+++ tuple.h
@@ -4,12 +4,12 @@
 #include "params.h"
 
 typedef struct {
-  uint16_t d;
-  uint16_t a;
-  uint16_t b;
-  uint16_t d1;
-  uint16_t a1;
-  uint16_t b1;
+  uint32_t d;
+  uint32_t a;
+  uint32_t b;
+  uint32_t d1;
+  uint32_t a1;
+  uint32_t b1;
 } tuple;
 
 tuple gen_tuple(uint32_t X, params *P);
